requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
jinja2==3.1.5
pydantic==2.10.5
fpdf2==2.8.2
python-dotenv==1.0.1
langchain==0.3.13
langchain-groq==0.2.2
langchain-community==0.3.13
langgraph==0.2.60
pypdf==5.1.0
docx2txt==0.8
unstructured[docx]


